bsp: fomu: fix rom address for example programs
During development, the ROM address was at offset 0x2001a000
(due to the fact that the test program was located immediately
following the bitstream).

In normal Fomus, the ROM address is at offset 0x20040000, in order
to take into account additional bitstreams and bootloaders.

Correct the address to the offset in order to get examples working
with existing Fomus.

Signed-off-by: Sean Cross <sean@xobs.io>